+++ routes/api/index.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 
 router.use('/users', require('./users'));
 router.use('/departments', require('./departments'));
-//router.use('/indicators', require('./indicators'));
+router.use('/indicators', require('./indicators'));
 router.use('/periods', require('./periods'));
 //router.use('/charts', require('./charts'));
 router.use('/auth', require('./authentication'));
